Use headings instead of numbered list for steps
I find headings more maintainable: it's now easy to insert/remove
steps in the guide without having to renumber everything in the
Markdown source (to avoid renumbering one could use "0." in the
source, but that looks bad when the source is read as-is).

Using headings also removes a layer of nesting and some GitHub
rendering problems due to the way they interpret numbered lists.