README: re-wrote build instructions, using pkgs
Completely re-wrote and simplified the build instructions. The new
instructions involve installation zerovm and zvm-validator from
packages, instead of installing manually.

NOTE: I removed the section which instructs the user to run zerovm tests
by invoking `./ftests.sh` from the ZEROVM_ROOT. We need to figure out a
good context in which to run these tests. They do require a working
toolchain to function, but they are part of the zerovm repo (which is a
prerequisite for building the toolchain), so there's kind of a catch 22
here that we need to think about. Perhaps it's best to move these tests
to the toolchain repo.